refactors. code style.
