+++ apps/editing-toolkit/editing-toolkit-plugin/wpcom-block-editor-nux/src/welcome-tour/style-tour.scss
@@ -66,7 +66,7 @@ $welcome-tour-button-background-color: #32373c; // former $dark-gray-700. TODO: 
 	}
 }
 
-.wpcom-editor-welcome-tour > .tour-kit__frame {
+.wpcom-editor-welcome-tour > .tour-kit-frame__container {
 	box-shadow: none;
 }
 
@@ -184,8 +184,8 @@ $welcome-tour-button-background-color: #32373c; // former $dark-gray-700. TODO: 
 			opacity: 0;
 		}
 
-		.tour-kit__frame:hover &,
-		.tour-kit__frame:focus-within & {
+		.tour-kit-frame__container:hover &,
+		.tour-kit-frame__container:focus-within & {
 			.components-button {
 				opacity: 0.7;
