+++ packages/tour-kit/src/components/tour-kit-portal.tsx
@@ -5,7 +5,7 @@ import { createPortal, useEffect, useRef } from '@wordpress/element';
 /**
  * Internal Dependencies
  */
-import TourFrame from './tour-frame';
+import TourKitFrame from './tour-kit-frame';
 import type { Config } from '../types';
 
 import '../styles.scss';
@@ -19,7 +19,7 @@ const TourPortal: React.FunctionComponent< Props > = ( { config } ) => {
 
 	useEffect( () => {
 		// @todo clk "tour-kit"
-		portalParent.classList.add( 'tour-kit__portal-parent' );
+		portalParent.classList.add( 'tour-kit-portal' );
 		document.body.appendChild( portalParent );
 
 		return () => {
@@ -27,7 +27,7 @@ const TourPortal: React.FunctionComponent< Props > = ( { config } ) => {
 		};
 	}, [ portalParent ] );
 
-	return <div>{ createPortal( <TourFrame config={ config } />, portalParent ) }</div>;
+	return <div>{ createPortal( <TourKitFrame config={ config } />, portalParent ) }</div>;
 };
 
 export default TourPortal;
